requirements.txt
Pillow>=5.3.0
aiohttp
cowpy
emoji
gTTS-token>=1.1.3
gTTS>=2.0.1
gitpython
google-api-python-client
oauth2client
google_images_download>=2.7.1
googletrans>=2.4.0
gtts>=2.0.1
hachoir
pybase64>=0.4.0
python-dotenv
pytube>=9.5.1
pytz
requests>=2.18.4
speedtest-cli>=2.0.2
telegraph
telethon>=1.10.5
wikipedia>=1.4.0
search-engine-parser
dnspython
lxml
humanize
pyDownload
psutil
pylast
numba
tqdm